[refactor] move db seed functions to its own script
