+++ server/db/helpers.js
@@ -1,7 +1,5 @@
 var db = require('./config.js');
-var async = require('async');
 var helpers = {};
-
 
 //---------------AUTHENTICATION----------------------
 
@@ -50,7 +48,6 @@ helpers.signupUser = function(userObj) {
 
 //-----------------USER PROFILE--------------------------
 
-
 helpers.updateUser = function(profileUpdateObj) {
   return db.User.findOne({
   	where: {'username': profileUpdateObj.username}
@@ -76,16 +73,14 @@ helpers.updateUser = function(profileUpdateObj) {
 };
 
 
-//------------------GET USERS-------------------------
+//------------------GET RECOMMENDED USERS-------------------------
 
 helpers.getAllUsers = function(username) {
   return db.User.findAll()
   .then(function(usersArray) {
     var result = helpers.getRecommendations(usersArray, username);
-    //call
     // console.log('HERE ARE ALL USERS', usersArray);
     return result;
-    // return usersArray;
   });
 };
 
@@ -115,93 +110,5 @@ helpers.getRecommendations = function(usersArray, username) {
 
   return usersArray;
 };
-//get user obj with that username from db
-//apply a filter to the list of all users based on some
-//property of that user obj
-
-//filter by skills to teach
-//some sort of intersection between arrays
-//we'll rank you higher than someone who only has two
-
-//api/users/username
-
-
-//--------------------FUNCTION TESTS
-// db.Skill.bulkCreate([
-// 	{	name: "AngularJS" },
-// 	{ name: "JavaScript" },
-// 	{ name: "Express"}
-// 	]).then(function() {
-// 		// return models.Skill.findAll();
-// 	})
-
-// db.User.create({
-//     username: "Tom123",
-//     name: "Tom Test",
-//     password: "abc123",
-//     email:"tom@tom.com",
-//     phoneNumber: '415-222-3215'
-//   }).then(function() {
-//   });
-
-// helpers.addMessage({
-// 	username: "Tom123",
-// 	recipientName: 'Rachel111',
-// 	text: 'Oh hello friend. howz it?!'
-// });
-
-// helpers.updateUserSkills({
-// 	username: 'Tom123',
-// 	toLearn: ['AngularJS'],
-// 	toTeach: ['JavaScript', 'Express']
-// }).then(function() {
-// 	console.log("hello success");
-// });
-
-helpers.seedDatabase = function() {
-
-  var array = [1,2,3,4,5,6,7,8,9,10];
-
-  async.each(array, function(i, next) {
-
-      //query right here
-      db.User.create({
-        username: 'user'+i,
-        password: 'password'+i,
-        email: 'user'+i+'@email.com',
-        name: 'user'+i,
-        karmaPoints: Math.floor(Math.random() * 6)
-        // toLearn: [],
-        // toTeach: []
-      }).then(function() {
-        console.log('User successfully created');
-        next();
-      });
-
-  }, function(err) {
-    if (err) {
-      console.log('An item failed to process');
-
-    } else {
-      console.log('successfully went through all of them');
-    }
-  });
-};
-
-
-// helpers.seedDatabase();
-
-
-// helpers.updateUserBasics({
-// 	username: 'Rachel111',
-// 	location: 'San Francisco',
-// 	email: 'rachel@rachelCity.com',
-// 	github: 'https://github.com/dearamerican',
-// 	summary: 'Hello I am a programmer now, woot!',
-// 	password: 'resetToSomethingHashedUp'
-// }).then(function(user) {
-// 	console.log("WOOT HERE IS USER", user);
-// });
 
 module.exports = helpers;
-
